requirements.txt
aiogram>=3.4,<4
asyncpg>=0.29,<1
aiohttp>=3.9,<4
python-dotenv>=1.0,<2
